// src/types/clinicalHistory.ts
export interface HeredofamiliaresCondition {
    has: 'SI' | 'NO' | '';
    parentescos: {
        padre: boolean;
        madre: boolean;
        hermanos: boolean;
        otros: boolean;
    },
    especifique: string;
}

export interface PathologicalCondition {
  has: 'SI' | 'NO' | '';
  especifique: string;
}